CapAndHomalg.jl

Introduction

The Julia package CapAndHomalg provides simplified access to the repositories of the GAP packages hosted on the GitHub organization homalg-project, most of which are based on the

CAP project,

homalg project.

These are open source GAP multi-package projects for constructive category theory and homological algebra with applications to module theory of commutative and non-commutative algebras and algebraic geometry.

Installation

Install Julia

To install Julia follow the first Steps 0 and 1 on our installation page.

Install CapAndHomalg.jl

Then start julia in a terminal and add the package CapAndHomalg:

julia> using Pkg; Pkg.add("CapAndHomalg")

julia> using CapAndHomalg

Testing

The correctness of the installation and the availability of the functionality can at any time be tested using

julia> using Pkg; Pkg.test("CapAndHomalg")

Updating

After each update of the Julia package GAP a rebuild is (probably) necessary:

julia> using Pkg; Pkg.update("CapAndHomalg")

julia> Pkg.build("CapAndHomalg")

This will also clone the repositories listed in CapAndHomalg.PACKAGES_TO_DOWNLOAD using DownloadAllPackagesFromHomalgProject and compile the packages listed in CapAndHomalg.PACKAGES_TO_COMPILE using GAP.Packages.install(pkgname).

CapAndHomalg.version — Constant

CapAndHomalg.version

The version number of the loaded CapAndHomalg. Please mention this number in any bug report.

CapAndHomalg.PKG_DIR — Constant

CapAndHomalg.PKG_DIR

The directory in which to install GAP packages from the GitHub organization https://github.com/homalg-project/. It is equal to joinpath(pathof(CapAndHomalg), "pkg").

CapAndHomalg.PACKAGES_TO_DOWNLOAD — Constant

CapAndHomalg.PACKAGES_TO_DOWNLOAD

List of packages which will be considered by

DownloadAllPackagesFromHomalgProject()

UpdateAllPackagesFromHomalgProject()

RemoveAllPackagesFromHomalgProject()

CapAndHomalg.DownloadAllPackagesFromHomalgProject — Function

DownloadAllPackagesFromHomalgProject()

Apply DownloadPackageFromHomalgProject to all packages listed in PACKAGES_TO_DOWNLOAD.

CapAndHomalg.UpdateAllPackagesFromHomalgProject — Function

UpdateAllPackagesFromHomalgProject()

Apply UpdatePackageFromHomalgProject to all packages listed in PACKAGES_TO_DOWNLOAD.

CapAndHomalg.RemoveAllPackagesFromHomalgProject — Function

RemoveAllPackagesFromHomalgProject()

Apply RemovePackageFromHomalgProject to all packages listed in PACKAGES_TO_DOWNLOAD.

CapAndHomalg.DownloadPackageFromHomalgProject — Function

DownloadPackageFromHomalgProject(pkgname)

Clone the repository named pkgname from the GitHub organization https://github.com/homalg-project/ to the subdirectory CapAndHomalg.PKG_DIR. On success return true and on failure false.

CapAndHomalg.UpdatePackageFromHomalgProject — Function

UpdatePackageFromHomalgProject(pkgname)

Update the repository named pkgname located in the subdirectory CapAndHomalg.PKG_DIR from the GitHub organization https://github.com/homalg-project/. If the package directory does not exist locally then invoke DownloadPackageFromHomalgProject(pkgname). On success return true and on failure false.

CapAndHomalg.RemovePackageFromHomalgProject — Function

RemovePackageFromHomalgProject(pkgname)

Delete the repository named pkgname from the subdirectory CapAndHomalg.PKG_DIR. On success return true and on failure false. Removing a repository and re-downloading it might be useful if udpating it fails.

CapAndHomalg.PACKAGES_TO_COMPILE — Constant

CapAndHomalg.PACKAGES_TO_COMPILE

The list of all GAP packages that will downloaded (once) and installed by GAP.Packages.install when using Pkg; Pkg.build("HomalgProject") or when CompilePackagesForHomalgProject is invoked. The latter should be called once GAP.jl gets updated.

Software dependency

CapAndHomalg relies on the

computer algebra systems	through the Julia packages

GAP	Gap.jl

Singular	Singular_jll.jl

all of which are components of the computer algebra system OSCAR.

Some of the bundled packages use the GAP packages

IO

ferret

json

QPA2

and the

third party software	through the GAP packages	and the Julia packages

Graphviz	Digraphs	Graphviz_jll.jl

4ti2	4ti2Interface	lib4ti2_jll.jl

<!– | cddlib | CddInterface | cddlib_jll.jl | | Normaliz | NormalizInterface | normaliz_jll.jl | –>

General Disclaimer

The software comes with absolutely no warranty and will most likely have errors. If you use it for computations, please check the correctness of the result very carefully.

This software is licensed under the LGPL, version 3, or any later version.

Funding

The development of this package and many of the GAP packages hosted on the GitHub organization homalg-project was partially funded by the DFG (German Research Foundation) through the

Special Priority Project SPP 1489,

Transregional Collaborative Research Centre SFB-TRR 195.

Index

CapAndHomalg.PACKAGES_TO_COMPILE

CapAndHomalg.PACKAGES_TO_DOWNLOAD

CapAndHomalg.PKG_DIR

CapAndHomalg.version

CapAndHomalg.DownloadAllPackagesFromHomalgProject

CapAndHomalg.DownloadPackageFromHomalgProject

CapAndHomalg.RemoveAllPackagesFromHomalgProject

CapAndHomalg.RemovePackageFromHomalgProject

CapAndHomalg.UpdateAllPackagesFromHomalgProject

CapAndHomalg.UpdatePackageFromHomalgProject